fix: make initGoogleDrive() actually authenticate with Google
Init previously only validated the SA JSON structure without ever
contacting Google. Now it signs a JWT and exchanges it at the token
endpoint — proving the credentials work, not just that the JSON parses.

Extracted verifyServiceAccountAuth() into app.ts so both init and the
test-connection endpoint share the same auth path.
